建筑结构复核 倾斜测量 两模块基础功能+数据存储

+++ app/src/main/java/com/sribs/bdd/v3/ui/check/ModuleCreateByTypeFloorActivity.kt
@@ -147,7 +147,7 @@ class ModuleCreateByTypeFloorActivity : BaseActivity(), IProjectContrast.IModule
 
             moduleCreateTypePresenter.createLocalModule(
                 this,
-                mLocalProjectId.toInt(),
+                mLocalProjectId,
                 mBuildingId,
                 mModuleId,
                 mRemoteId,
